Renamed files and changed no items message
