+++ src/components/Store/CheckOut.jsx
@@ -1,14 +1,13 @@
-import React, { useState, useEffect, useContext } from 'react';
+import React, { useState } from 'react';
 import Form from 'react-bootstrap/Form';
 import Alert from 'react-bootstrap/Alert';
 
-import { Container, Row, Col, Button } from 'react-bootstrap';
+import { Container, Col, Button } from 'react-bootstrap';
 
 import {purchaseItems} from '../../utils/store/store';
 
-import StoreContext from './context';
 
-const Purchase = () => {
+const CheckOut = () => {
 
   // For user feedback on API responses
   const [showAlert, setShowAlert] = useState(false);
@@ -16,30 +15,11 @@ const Purchase = () => {
   const [alertMsg, setAlertMessage] = useState('');
   const [isLoading, setLoading] = useState(false);
 
-  const { items, addCart } = useContext(StoreContext);
-
-  const [item, setItem] = useState({})
-  const [type, setType] = useState('')
-
   const [email, setEmail] = useState('');
   const [address, setAddress] = useState('');
   const [apt, setApt] = useState('');
   const [city, setCity] = useState('');
   const [zip, setZip] = useState('');
-
-  const refresh = () => {
-    if(items.length !== 0){
-      setItem(items.find((elm) => elm.item_id === 1));
-    }
-  }
-
-  useEffect(() => {
-    setType(item.types && Array.isArray(item.types) ? item.types[0] : '');
-  }, [item]);
-
-  useEffect(() => {
-    refresh();
-  }, [items])
 
   const handleSubmit = async () => {
     setLoading(true);
@@ -57,53 +37,9 @@ const Purchase = () => {
     }
   };
 
-  const handleCart = () => {
-    const obj = { ...item};
-    obj.type = type;
-    addCart(obj);
-  }
-
 
   return (
     <div>
-      <Container>
-        <Row>
-          <Col md={4}>
-            <img src={item.picture} alt={item.name} />
-          </Col>
-          <Col md={8}>
-            <div className="py-4">
-              <h2>{item.name}</h2>
-              <h3>{item.price}</h3>
-              <Form>
-                <Form.Group controlId="formGridState">
-                  <Form.Label>Type</Form.Label>
-                  <Form.Control
-                    as="select" 
-                    onChange={(e) => {setType(e.target.value)}}
-                  >
-                    {item.types && item.types.map(elm => {
-                    return (<option key={elm}>{elm}</option>)
-                  })}
-                  </Form.Control>
-                </Form.Group>
-                <div className="py-2">
-                  <Button
-                    onClick={() => {handleCart()}}
-                  >
-                    Add to cart
-                  </Button>
-                </div>
-                <div>
-                  <Button>
-                    Purchase
-                  </Button>
-                </div>
-              </Form>
-            </div>
-          </Col>
-        </Row>
-      </Container>
       
       <Container>
         <Form autoComplete="off">
@@ -197,4 +133,4 @@ const Purchase = () => {
   );
 };
 
-export default Purchase;
+export default CheckOut;
